Misc improvements

+++ packages/cli/src/commands/createApp.ts
@@ -22,50 +22,7 @@ function createPackageJson(root: string, appName: string): void {
   fs.writeFileSync(path.join(root, 'package.json'), JSON.stringify(packageJson, null, 2) + os.EOL);
 }
 
-function makeCaretRange(dependencies: any, name: string) {
-  const version = dependencies[name];
-
-  if (typeof version === 'undefined') {
-    console.error(chalk.red(`Missing ${name} dependency in package.json`));
-    process.exit(1);
-  }
-
-  let patchedVersion = `^${version}`;
-
-  if (!semver.validRange(patchedVersion)) {
-    console.error(
-      `Unable to patch ${name} dependency version because version ${chalk.red(version)} will become invalid ${chalk.red(
-        patchedVersion,
-      )}`,
-    );
-    patchedVersion = version;
-  }
-
-  dependencies[name] = patchedVersion;
-}
-
-function setCaretRangeForRuntimeDeps(packageName: string) {
-  const packagePath = path.join(process.cwd(), 'package.json');
-  const packageJson = require(packagePath);
-
-  if (typeof packageJson.dependencies === 'undefined') {
-    console.error(chalk.red('Missing dependencies in package.json'));
-    process.exit(1);
-  }
-
-  const packageVersion = packageJson.dependencies[packageName];
-  if (typeof packageVersion === 'undefined') {
-    console.error(chalk.red(`Unable to find ${packageName} in package.json`));
-    process.exit(1);
-  }
-
-  makeCaretRange(packageJson.dependencies, 'react');
-  makeCaretRange(packageJson.dependencies, 'react-dom');
-
-  fs.writeFileSync(packagePath, JSON.stringify(packageJson, null, 2) + os.EOL);
-}
-
-function abort(reason: any, root: string, appName: string) {
+function abort(reason: any, root: string, appName: string, name: string) {
   console.log();
   console.log('Aborting installation.');
   if (reason.command) {
@@ -81,7 +38,6 @@ function abort(reason: any, root: string, appName: string) {
   const currentFiles = fs.readdirSync(path.join(root));
   currentFiles.forEach(file => {
     knownGeneratedFiles.forEach(fileToMatch => {
-      // This remove all of knownGeneratedFiles.
       if (file === fileToMatch) {
         console.log(`Deleting generated file... ${chalk.cyan(file)}`);
         fs.removeSync(path.join(root, file));
@@ -89,8 +45,7 @@ function abort(reason: any, root: string, appName: string) {
     });
   });
   const remainingFiles = fs.readdirSync(path.join(root));
-  if (!remainingFiles.length) {
-    // Delete target folder if empty
+  if (name && !remainingFiles.length) {
     console.log(`Deleting ${chalk.cyan(`${appName}/`)} from ${chalk.cyan(path.resolve(root, '..'))}`);
     process.chdir(path.resolve(root, '..'));
     fs.removeSync(path.join(root));
@@ -119,25 +74,24 @@ export default function createApp(name: string, verbose: boolean, version: strin
   guardNpmVersion(isYarn);
 
   const solonScriptsToInstall = getSolonScriptsToInstall(version, originalDirectory);
-  const allDependencies = ['react', 'react-dom', solonScriptsToInstall];
+  const allDependencies = [solonScriptsToInstall];
 
   console.log('Installing packages. This might take a couple of minutes.');
   getSolonScriptsPackageName(solonScriptsToInstall)
     .then(packageName => checkIfOnline(isYarn).then(isOnline => ({ isOnline, packageName })))
     .then(info => {
       const { isOnline, packageName } = info;
-      console.log(`Installing ${chalk.cyan('react')}, ${chalk.cyan('react-dom')}, and ${chalk.cyan(packageName)}...`);
+      console.log(`Installing ${chalk.cyan(packageName)}...`);
       console.log();
 
       return installDependencies(root, isYarn, allDependencies, verbose, isOnline).then(() => packageName);
     })
     .then(packageName => {
       guardNodeVersion(packageName);
-      setCaretRangeForRuntimeDeps(packageName);
 
-      const scriptsPath = path.resolve(process.cwd(), 'node_modules', packageName, 'scripts', 'init.js');
+      const scriptsPath = path.resolve(process.cwd(), 'node_modules', packageName, 'lib', 'commands', 'init.js');
       const init = require(scriptsPath);
-      init(root, appName, verbose, originalDirectory);
+      // init(root, appName, verbose, originalDirectory);
     })
-    .catch(reason => abort(reason, root, appName));
+    .catch(reason => abort(reason, root, appName, name));
 }
